requirements.txt
# nonebot_plugin_osumania_toolkit 依赖文件
# 2026-03-13

nonebot2>=2.3.0
nonebot-adapter-onebot>=2.0.0
nonebot-plugin-localstore>=0.3.0
aiohttp>=3.8.0
Pillow>=10.0.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pydantic>=2.0.0
pandas>=3.0.0